fix: "pause" deprecated in Strimzi 38.0
The "pause" key in thet config spec is deprecated in Strimzi 38.0. It
has been replaced by "state". In order to duplicate what we are
currently doing with the operator, we need to change this to `state:
running`.

This is required when moving to amq-streams 2.6 and Kafka 3.5.
Otherwise, the connectors will continously fall over

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>